Block VM if all tasks are sleeping
We were previously running with a busy loop while waiting for sleeping
tasks to be runnable. That means that simple programs like the hello
example (which prints the current time every 5 seconds) could use almost
100% CPU while running.

Now if all tasks are sleeping, we schedule a timer event for the next
time a task is runnable, and then block to wait for the event. The hello
program now uses effectively 0% CPU.

Resolves #12.