Add timeInfo.checkSystemTime and checkBootTime
This cleans up and makes explicit time synchronization in the shadow
time struct. Now the different time values (boot and system) are updated
independently, based on need; that is, boot time is not checked when
getting a monotonic clock value.

The atomic version synchronization routine is extracted into its own
method, which functions as a template for both clock update functions.