crypto/rand: add insecure reader implementation
crypto/rand.Reader was previously nil, causing users of the rand package
to segfault.

We don't have yet everything in place to provide a source of
cryptographically secure random data, so for now we leverage the
insecure random generator in math/rand.

Follow https://github.com/atmanos/atmanos/issues/28 for long-term
solutions to this.